Move find run command functions to cli utils
